feat(editor): documenten via TrajectMenu-sheet i.p.v. pane-panel, NLDD-only
Vervangt het losse 'Documenten'-pane-panel door dezelfde flow als 'Nieuw
traject': een 'Documenten…'-item in het TrajectMenu opent een sheet met de
file-lijst + aanmaken/verwijderen; een file kiezen/aanmaken opent 'm in het
modeless/movable nldd-window.

- Nieuw: components/TrajectDocuments.vue (sheet + window + delete-modal),
  volledig opgebouwd uit NLDD-componenten, geen custom CSS:
  nldd-list/list-item + nldd-button (file-lijst), nldd-text-field (aanmaken),
  nldd-multi-line-text-field (editor), nldd-rich-text (markdown-preview),
  nldd-inline-dialog (alle banners), nldd-divider/spacer (layout).
- Nieuw: composables/useDocumentsSheet.js — singleton open-state zodat het
  (meervoudig gemounte) TrajectMenu het één keer gerenderde component opent.
- TrajectMenu: 'Documenten…'-item (bij actief traject).
- EditorApp + LibraryApp: TrajectDocuments één keer gemount; documenten-
  pane-view, panel.documents-flag (frontend + backend) en DocumentsPanel.vue
  verwijderd.
