+++ frontend/src/composables/useFeatureFlags.js
@@ -14,10 +14,6 @@ const DEFAULTS = {
   // Notes pane (RFC-005/RFC-018). Off by default until the feature is past
   // the display-only MVP and the corpus has notes for more than one law.
   'panel.notes': false,
-  // Documents panel (traject-scoped markdown/text files). On by default
-  // like the other panels; availableViews only surfaces it when a traject
-  // is active, and the panel toggle is likewise only offered with a traject.
-  'panel.documents': true,
   // Note authoring (RFC-018 write path, MVP: localStorage + manual export).
   // Off by default: it is a separate, opt-in capability on top of viewing.
   'notes.create': false,
